refactor(constants): move constants from src/utils to src
