+++ src/data_gen/generate.py
@@ -7,10 +7,10 @@ from typing import TYPE_CHECKING
 from datasets import Dataset, DatasetDict, concatenate_datasets, load_dataset
 from structlog import getLogger
 
+from src import constants as const
 from src.data_gen.reasoner import reason
 from src.data_prep.data_hf import load_txt_by_lines, push_data_to_hf
 from src.data_prep.dedup import decontaminate, deduplicate
-from src.utils import constants as const
 from src.utils.common import read_prompt
 
 if TYPE_CHECKING:
